# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Medical Appointment API
  version: 1.0.0
  description: API documentation for Medical Appointment System
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - name
        - email
        - password
        - phoneNumber
        - CIN
        - role
      properties:
        name:
          type: string
          example: "Dr. John Smith"
        email:
          type: string
          format: email
          example: "john.smith@example.com"
        password:
          type: string
          format: password
          example: "Doctor123!"
        phoneNumber:
          type: string
          example: "+1234567890"
        CIN:
          type: string
          example: "D123456"
        role:
          type: string
          enum: [Doctor, Patient, Admin]
        specialization:
          type: string
          example: "Cardiology"
        department:
          type: string
          format: uuid
        profileImage:
          type: string
          format: uri
        diplomaImage:
          type: string
          format: uri
        availability:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
                enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
              slots:
                type: array
                items:
                  type: object
                  properties:
                    time:
                      type: string
                      example: "09:00"
                    isBooked:
                      type: boolean
                      default: false
        isValidated:
          type: boolean
          default: false
        medicalHistory:
          type: string
    Department:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: "Cardiology"
        description:
          type: string
          example: "Diagnosis and treatment of heart diseases"
        doctors:
          type: array
          items:
            type: string
            format: uuid
...
paths:
...
  /api/doctors:
    get:
      tags: [Doctors]
      summary: Get all doctors
      responses:
        200:
          description: List of all doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    
    post:
      tags: [Doctors]
      summary: Register new doctor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                phoneNumber:
                  type: string
                CIN:
                  type: string
                specialization:
                  type: string
                department:
                  type: string
                availability:
                  type: string
                profileImage:
                  type: string
                  format: binary
                diplomaImage:
                  type: string
                  format: binary
      responses:
        201:
          description: Doctor registered successfully
  /api/doctors/{id}:
    get:
      tags: [Doctors]
      summary: Get doctor by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    
    put:
      tags: [Doctors]
      summary: Update doctor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                specialty:
                  type: string
                availableTimes:
                  type: string
                profileImage:
                  type: string
                  format: binary
                diplomaImage:
                  type: string
                  format: binary
      responses:
        200:
          description: Doctor updated successfully

    delete:
      tags: [Doctors]
      summary: Delete doctor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Doctor deleted successfully
...
  /api/departments:
    get:
      tags: [Departments]
      summary: Get all departments
      responses:
        200:
          description: List of all departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
    
    post:
      tags: [Departments]
      summary: Create department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        201:
          description: Department created successfully
...
  /api/public/departments/{departmentId}/doctors:
    get:
      tags: [Public]
      summary: Get validated doctors by department (public)
      parameters:
        - in: path
          name: departmentId
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of validated doctors in the department
...
security:
  - bearerAuth: [] 